+++ src/main/java/vampire/city/repositories/AvatarRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import vampire.city.model.Avatar;
 
-public interface AvatarRepository extends JpaRepository<Avatar, Long> {
+public interface AvatarRepository extends JpaRepository<Avatar, Integer> {
     
 }
